fixed WG api hit count
Wunderground api call tracker was not correctly rembering the number of
api calls that had been made in a day.  Hacked it to work by declaring
the variable public and writeing the value retrieved from the settings
file directly.  previously I was passing it to a method that assigned
the value but for whatever reason that wasnt working...as a lazy man I
left the method in place.